+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - `jxl-oxide`: Accept `u8` and `u16` output buffers in `ImageStream::write_to_buffer` (#366).
+- `jxl-oxide`: Add `image` integration under a feature flag (#368).
 
 ### Changed
 - `jxl-color`: Use better PQ to HLG method (#348).